Revise TextIndexer with test
